[feat] Mypage, Cafeteria Fragment 로 전환 후 MainActivity에 있어야 하는 로직 옮기기. activityviewmodel 사용을 위한 라이브러리 추가 및 프래그먼트용 토스트 확장 함수 생성.
