+++ app/src/main/res/layout/fragment_cafeteria.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/gray100"
-    tools:context=".presentation.main.CafeteriaActivity">
+    tools:context=".presentation.main.cafeteria.CafeteriaFragment">
 
     <!--  로고 및 상단 바  -->
     <androidx.constraintlayout.widget.ConstraintLayout
@@ -154,23 +154,5 @@
             app:layout_constraintBottom_toTopOf="@+id/layout_bottom_navi"
             />
     </LinearLayout>
-
-    <!-- 바텀 네비게이션 바 -->
-    <com.google.android.material.bottomnavigation.BottomNavigationView
-        android:id = "@+id/bottom_navi_bar"
-        style="@style/Caption2"
-        android:layout_width="match_parent"
-        android:layout_height="wrap_content"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        android:paddingVertical="15dp"
-        android:elevation="3dp"
-        android:background="@drawable/shape_corner_top"
-        app:itemRippleColor="@color/white"
-        android:textAlignment="center"
-        app:itemTextColor="@color/selector_bottom_navi_item"
-        app:itemIconTint="@color/selector_bottom_navi_item"
-        app:menu="@menu/menu_bottom_navigation" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
